Move code to misc.go Rename util_test.go to misc_test.go. Add some unitary test code.
