+++ utils/misc_test.go
@@ -79,3 +79,23 @@ func TestDependencies(t *testing.T) {
 	assert.Equal(t, "http://github.com/user/repo", record.BaseRepo, "BaseRepo is wrong")
 	assert.Equal(t, "/subfolder1/subfolder2", record.SubFolder, "SubFolder is wrong")
 }
+
+func TestParseOpenWhisk(t *testing.T) {
+	openwhiskHost := "https://openwhisk.ng.bluemix.net"
+    openwhisk, err := ParseOpenWhisk(openwhiskHost)
+    assert.Equal(t, nil, err, "parse openwhisk info error happened.")
+    converted := ConvertToMap(openwhisk)
+    assert.Equal(t, 1, len(converted["nodejs"]), "not expected length")
+    assert.Equal(t, 1, len(converted["php"]),  "not expected length")
+    assert.Equal(t, 1, len(converted["java"]), "not expected length")
+    assert.Equal(t, 3, len(converted["python"]), "not expected length")
+    assert.Equal(t, 2, len(converted["swift"]), "not expected length")
+}
+
+func TestNewZipWritter(t *testing.T) {
+    filePath := "../tests/src/integration/zipaction/actions/cat"
+    zipName := filePath + ".zip"
+    err := NewZipWritter(filePath, zipName).Zip()
+    defer os.Remove(zipName)
+    assert.Equal(t, nil, err, "zip folder error happened.")
+}
